Install comskip-gui only if it was built

+++ PKGBUILD
@@ -1,7 +1,7 @@
 # Maintainer: Nick Burrett <nick@sqrt.co.uk>
 pkgname=comskip
 pkgver=0.81.095
-pkgrel=2
+pkgrel=3
 epoch=1
 pkgdesc='Comskip is a free MPEG commercial break detector'.
 _gitname='comskip'
@@ -33,7 +33,9 @@ build() {
 package() {
   mkdir -p "$pkgdir/usr/bin"
   install -m755 "${srcdir}/${_gitname}/comskip" "$pkgdir/usr/bin/comskip"
-  install -m755 "${srcdir}/${_gitname}/comskip-gui" "$pkgdir/usr/bin/comskip-gui"
+  if [ -e "${srcdir}/${_gitname}/comskip-gui" ]; then
+    install -m755 "${srcdir}/${_gitname}/comskip-gui" "$pkgdir/usr/bin/comskip-gui"
+  fi
 }
 
 # vim:set ts=2 sw=2 et:
